+++ src/rpn.py
@@ -145,11 +145,11 @@ class RPN:
             Returns:
                 clipped_boxes (tf.Tensor): The clipped absolute coordinates in the feature maps.
         '''
-
-        x1 = tf.clip_by_value(boxes[:, :, 0], 0, fm_shape[1] - 1)
-        y1 = tf.clip_by_value(boxes[:, :, 1], 0, fm_shape[2] - 1)
-        x2 = tf.clip_by_value(boxes[:, :, 2], 0, fm_shape[1] - 1)
-        y2 = tf.clip_by_value(boxes[:, :, 3], 0, fm_shape[2] - 1)
+        offset = self.config.rpn_clip_offset
+        x1 = tf.clip_by_value(boxes[:, :, 0], 0, fm_shape[1] - 1 - offset)
+        y1 = tf.clip_by_value(boxes[:, :, 1], 0, fm_shape[2] - 1 - offset)
+        x2 = tf.clip_by_value(boxes[:, :, 2], 0, fm_shape[1] - 1 - offset)
+        y2 = tf.clip_by_value(boxes[:, :, 3], 0, fm_shape[2] - 1 - offset)
 
         clipped_boxes = tf.stack([x1, y1, x2, y2], axis=-1)
         return clipped_boxes